# apps/api/docs/swagger.yaml
definitions:
  dto.AIChatRequest:
    properties:
      domain:
        type: string
      message:
        type: string
    required:
    - message
    type: object
  dto.AIQuestionRequest:
    properties:
      domain:
        type: string
      topic:
        type: string
    required:
    - domain
    - topic
    type: object
...
  dto.ListeningLesson:
    properties:
      audio_url:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  dto.ListeningSubmissionRequest:
    properties:
      answers:
        items:
          type: string
        type: array
    required:
    - answers
    type: object
...
  dto.PronunciationRequest:
    properties:
      text:
        type: string
    required:
    - text
    type: object
  dto.PronunciationResult:
    properties:
      accuracy:
        type: number
      feedback:
        type: string
    type: object
...
  dto.SpeakingSession:
    properties:
      id:
        type: string
      started_at:
        type: string
      status:
        type: string
      stopped_at:
        type: string
    type: object
...
  dto.UpdateWordRequest:
    properties:
      definition:
        type: string
      example:
        type: string
      level:
        type: string
      part_of_speech:
        type: string
      phonetic:
        type: string
      word:
        type: string
    required:
    - definition
    - word
    type: object
...
  dto.VocabularyHistoryItem:
    properties:
      id:
        type: string
      result:
        type: string
      reviewed_at:
        type: string
      word_id:
        type: string
    type: object
...
  dto.WritingEvaluationRequest:
    properties:
      prompt:
        type: string
      text:
        type: string
    required:
    - prompt
    - text
    type: object
...
  model.VocabularyWord:
    properties:
      created_at:
        type: string
      definition:
        type: string
      example:
        type: string
      id:
        type: string
      level:
        type: string
      part_of_speech:
        type: string
      phonetic:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      word:
        type: string
    type: object
...
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      request_id:
        type: string
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: integer
      fields:
        additionalProperties:
          type: string
        description: validation field errors
        type: object
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /ai/chat:
    post:
      consumes:
      - application/json
      parameters:
      - description: AI chat payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AIChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Chat with AI coach
      tags:
      - ai
...
  /ai/evaluate-speaking:
    post:
      consumes:
      - application/json
      parameters:
      - description: AI speaking evaluation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AIChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Evaluate speaking with AI
      tags:
      - ai
  /ai/evaluate-writing:
    post:
      consumes:
      - application/json
      parameters:
      - description: AI writing evaluation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.WritingEvaluationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Evaluate writing with AI
      tags:
      - ai
  /ai/generate-question:
    post:
      consumes:
      - application/json
      parameters:
      - description: AI question payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AIQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Generate question with AI
      tags:
      - ai
...
  /listening/{id}:
    get:
      parameters:
      - description: Listening lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListeningLesson'
              type: object
      security:
      - BearerAuth: []
      summary: Get listening lesson by id
      tags:
      - listening
  /listening/{id}/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Listening lesson ID
        in: path
        name: id
        required: true
        type: string
      - description: Listening submission payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ListeningSubmissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Submit listening answers
      tags:
      - listening
  /listening/lessons:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search title, description or transcript
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ListeningLesson'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List listening lessons
      tags:
      - listening
...
  /speaking/pronunciation:
    post:
      consumes:
      - application/json
      parameters:
      - description: Pronunciation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PronunciationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.PronunciationResult'
              type: object
      security:
      - BearerAuth: []
      summary: Analyze pronunciation
      tags:
      - speaking
  /speaking/session/{id}:
    get:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.SpeakingSession'
              type: object
      security:
      - BearerAuth: []
      summary: Get speaking session by id
      tags:
      - speaking
  /speaking/session/start:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.SpeakingSession'
              type: object
      security:
      - BearerAuth: []
      summary: Start speaking session
      tags:
      - speaking
  /speaking/session/stop:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.SpeakingSession'
              type: object
      security:
      - BearerAuth: []
      summary: Stop speaking session
      tags:
      - speaking
...
  /vocabulary/history:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by word or definition
        in: query
        name: q
        type: string
      - description: Filter by result
        in: query
        name: result
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.VocabularyHistoryItem'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get vocabulary review history
      tags:
      - vocabulary
...
  /vocabulary/words/{id}:
    delete:
      parameters:
      - description: Word ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete vocabulary word
      tags:
      - vocabulary
    put:
      consumes:
      - application/json
      parameters:
      - description: Word ID
        in: path
        name: id
        required: true
        type: string
      - description: Word payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Update vocabulary word
      tags:
      - vocabulary
  /vocabulary/words/{wordId}:
    get:
      parameters:
      - description: Word ID
        in: path
        name: wordId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.VocabularyWord'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get a vocabulary word by id
      tags:
      - vocabulary
  /writing/{id}:
    get:
      parameters:
      - description: Writing ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get writing item by id
      tags:
      - writing
  /writing/evaluate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Writing evaluation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.WritingEvaluationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Evaluate writing text
      tags:
      - writing
...
swagger: "2.0"